Check for body missing @JsonObject annotation. (#25)
Not sure if it'd be better to do a separate detector for this? Also
@Body are not checking for @Json annotations - which is a difference
from the rest of the checks. Maybe that check should be added?

I still felt that adding this is a step towards better coverage for the
lint checks - even if not checking for @Json annotations currently.